requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.1
pydantic==2.5.0
sqlalchemy==2.0.23
PyPDF2==3.0.1
python-multipart==0.0.6
aiofiles==23.2.1
